latest i can do for permission

+++ app/Http/Controllers/ClientController.php
@@ -4,19 +4,30 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Client;
+use App\Models\PermissionRoleModel;
+use Illuminate\Support\Facades\Auth;
 
 class ClientController extends Controller
 {
     public function index()
-    {
-        $getClient = Client::orderBy('id', 'desc')->get();
-        return view('panel.client.list', compact('getClient'));
-    }
+{
+    $getClient = Client::orderBy('id', 'desc')->get();
 
-    public function add()
-    {
-        return view('panel.client.add');
+    $PermissionAdd = PermissionRoleModel::getPermission('Add Client', Auth::user()->role_id);
+    $PermissionEdit = PermissionRoleModel::getPermission('Edit Client', Auth::user()->role_id);
+    $PermissionDelete = PermissionRoleModel::getPermission('Delete Client', Auth::user()->role_id);
+
+    return view('panel.client.list', compact('getClient', 'PermissionAdd', 'PermissionEdit', 'PermissionDelete'));
+}
+
+public function add()
+{
+    $PermissionAdd = PermissionRoleModel::getPermission('Add Client', Auth::user()->role_id);
+    if (empty($PermissionAdd)) {
+        abort(404);
     }
+    return view('panel.client.add');
+}
 
     public function insert(Request $request)
     {
@@ -33,6 +44,10 @@ class ClientController extends Controller
 
     public function edit($id)
     {
+        $PermissionEdit = PermissionRoleModel::getPermission('Edit Client', Auth::user()->role_id);
+    if (empty($PermissionEdit)) {
+        abort(404);
+    }
         $getRecord = Client::findOrFail($id);
         return view('panel.client.edit', compact('getRecord'));
     }
@@ -52,6 +67,10 @@ class ClientController extends Controller
 
     public function delete($id)
     {
+        $PermissionDelete = PermissionRoleModel::getPermission('Delete Client', Auth::user()->role_id);
+    if (empty($PermissionDelete)) {
+        abort(404);
+    }
         $client = Client::findOrFail($id);
         $client->delete();
